Grey text on a white background is not very readable
Changed the white message background to the dark grey used on the sidebar, this makes the messages much more readable and does not affect messages which wish to have white backgrounds. I hate white and try my best to eliminate it as much as possible, it doesn't make sense for this theme.